Move track actions into common
