+++ src/containers/history-page/HistoryPageProvider.tsx
@@ -14,12 +14,12 @@ import {
 import { ID, UID } from 'common/models/Identifiers'
 import Status from 'common/models/Status'
 import { getUserId } from 'common/store/account/selectors'
+import * as socialActions from 'common/store/social/tracks/actions'
 import { getHistoryTracksLineup } from 'containers/history-page/store/selectors'
 import { useRecord, make } from 'store/analytics/actions'
 import { makeGetTableMetadatas } from 'store/lineup/selectors'
 import { getPlaying, getBuffering } from 'store/player/selectors'
 import { makeGetCurrent } from 'store/queue/selectors'
-import * as socialActions from 'store/social/tracks/actions'
 import { AppState } from 'store/types'
 import { profilePage } from 'utils/route'
 import { withNullGuard } from 'utils/withNullGuard'
