+++ src/containers/now-playing/NowPlaying.tsx
@@ -16,6 +16,13 @@ import {
 import { ID } from 'common/models/Identifiers'
 import { SquareSizes } from 'common/models/ImageSizes'
 import { getUserId } from 'common/store/account/selectors'
+import {
+  saveTrack,
+  unsaveTrack,
+  repostTrack,
+  undoRepostTrack,
+  shareTrack
+} from 'common/store/social/tracks/actions'
 import { open } from 'common/store/ui/mobile-overflow-menu/actions'
 import {
   OverflowAction,
@@ -49,13 +56,6 @@ import { AudioState } from 'store/player/types'
 import { makeGetCurrent } from 'store/queue/selectors'
 import { next, pause, play, previous, repeat, shuffle } from 'store/queue/slice'
 import { RepeatMode } from 'store/queue/types'
-import {
-  saveTrack,
-  unsaveTrack,
-  repostTrack,
-  undoRepostTrack,
-  shareTrack
-} from 'store/social/tracks/actions'
 import { AppState } from 'store/types'
 import { Genre } from 'utils/genres'
 import { pushUniqueRoute as pushRoute, profilePage } from 'utils/route'
