+++ src/containers/saved-page/SavedPageProvider.tsx
@@ -14,6 +14,7 @@ import {
 import { ID, UID } from 'common/models/Identifiers'
 import * as accountActions from 'common/store/account/reducer'
 import { getAccountWithSavedPlaylistsAndAlbums } from 'common/store/account/selectors'
+import * as socialActions from 'common/store/social/tracks/actions'
 import { updatePlaylistLastViewedAt } from 'containers/notification/store/actions'
 import { getPlaylistUpdates } from 'containers/notification/store/selectors'
 import { getSavedTracksLineup } from 'containers/saved-page/store/selectors'
@@ -21,7 +22,6 @@ import { TrackEvent, make } from 'store/analytics/actions'
 import { makeGetTableMetadatas } from 'store/lineup/selectors'
 import { getPlaying, getBuffering } from 'store/player/selectors'
 import { makeGetCurrent } from 'store/queue/selectors'
-import * as socialActions from 'store/social/tracks/actions'
 import { AppState } from 'store/types'
 import { isMobile } from 'utils/clientUtil'
 import { formatCount } from 'utils/formatUtil'
