+++ src/containers/saved-page/store/lineups/tracks/sagas.js
@@ -4,6 +4,7 @@ import { call, select, put, takeEvery } from 'redux-saga/effects'
 import Kind from 'common/models/Kind'
 import { getTracks as getCacheTracks } from 'common/store/cache/tracks/selectors'
 import { retrieveTracks } from 'common/store/cache/tracks/utils'
+import { SAVE_TRACK, UNSAVE_TRACK } from 'common/store/social/tracks/actions'
 import { makeUid } from 'common/utils/uid'
 import * as saveActions from 'containers/saved-page/store/actions'
 import {
@@ -15,7 +16,6 @@ import {
 import { LineupSagas } from 'store/lineup/sagas'
 import { getUid as getPlayerUid } from 'store/player/selectors'
 import * as queueActions from 'store/queue/slice'
-import { SAVE_TRACK, UNSAVE_TRACK } from 'store/social/tracks/actions'
 
 import { PREFIX, tracksActions as savedTracksActions } from './actions'
 const getSavedTracks = state => state.saved.tracks
