+++ src/containers/track-overflow-modal/ConnectedMobileOverflowModal.tsx
@@ -15,6 +15,13 @@ import { publishPlaylist } from 'common/store/cache/collections/actions'
 import { getCollection } from 'common/store/cache/collections/selectors'
 import { getTrack } from 'common/store/cache/tracks/selectors'
 import { getUser } from 'common/store/cache/users/selectors'
+import {
+  repostTrack,
+  undoRepostTrack,
+  saveTrack,
+  unsaveTrack,
+  shareTrack
+} from 'common/store/social/tracks/actions'
 import { close } from 'common/store/ui/mobile-overflow-menu/actions'
 import { getMobileOverflowModal } from 'common/store/ui/mobile-overflow-menu/selectors'
 import { OverflowSource } from 'common/store/ui/mobile-overflow-menu/types'
@@ -38,13 +45,6 @@ import {
   unsaveCollection,
   shareCollection
 } from 'store/social/collections/actions'
-import {
-  repostTrack,
-  undoRepostTrack,
-  saveTrack,
-  unsaveTrack,
-  shareTrack
-} from 'store/social/tracks/actions'
 import { followUser, unfollowUser, shareUser } from 'store/social/users/actions'
 import { AppState } from 'store/types'
 import { profilePage, playlistPage, albumPage } from 'utils/route'
