+++ src/containers/track-page/TrackPageProvider.tsx
@@ -18,6 +18,7 @@ import Status from 'common/models/Status'
 import Track from 'common/models/Track'
 import { getUserId } from 'common/store/account/selectors'
 import * as cacheTrackActions from 'common/store/cache/tracks/actions'
+import * as socialTracksActions from 'common/store/social/tracks/actions'
 import { open } from 'common/store/ui/mobile-overflow-menu/actions'
 import {
   OverflowAction,
@@ -52,7 +53,6 @@ import { makeGetLineupMetadatas } from 'store/lineup/selectors'
 import { getPlaying, getBuffering } from 'store/player/selectors'
 import { makeGetCurrent } from 'store/queue/selectors'
 import { getLocationPathname } from 'store/routing/selectors'
-import * as socialTracksActions from 'store/social/tracks/actions'
 import * as socialUsersActions from 'store/social/users/actions'
 import { AppState } from 'store/types'
 import { isMobile } from 'utils/clientUtil'
