+++ src/store/social/tracks/sagas.ts
@@ -11,6 +11,7 @@ import * as cacheActions from 'common/store/cache/actions'
 import { getTrack, getTracks } from 'common/store/cache/tracks/selectors'
 import { adjustUserField } from 'common/store/cache/users/sagas'
 import { getUser } from 'common/store/cache/users/selectors'
+import * as socialActions from 'common/store/social/tracks/actions'
 import { makeKindId } from 'common/utils/uid'
 import * as signOnActions from 'containers/sign-on/store/actions'
 import AudiusBackend from 'services/AudiusBackend'
@@ -19,7 +20,6 @@ import { make } from 'store/analytics/actions'
 import { waitForBackendSetup } from 'store/backend/sagas'
 import * as confirmerActions from 'store/confirmer/actions'
 import { confirmTransaction } from 'store/confirmer/sagas'
-import * as socialActions from 'store/social/tracks/actions'
 import { formatShareText } from 'utils/formatUtil'
 import { waitForValue } from 'utils/sagaHelpers'
 import { share } from 'utils/share'
